requirements.txt
reflex==0.6.3
requests
uagents
python-dotenv
google.generativeai
cartesia
ffmpeg-python
websockets>=12.0